main-backend/requirements.txt
Django==4.1
djangorestframework==3.14.0
numpy==1.26.1
pandas==2.1.3
psycopg2==2.9.9
python-dotenv==1.0.0
requests==2.31.0